requirement.txt
python-dotenv
pandas
numpy
scikit-learn
lightgbm
feather-format
seaborn
xgboost
optbinning